Add host_software_version() and flush()
RemoteObject::host_software_version() returns the version of the c++ host
object, while RemoteObject::flush() clears the Serial buffer. The
SimpleSerial class now flushes it's buffer as soon as it connects. Before
making this change, it was possible for Serial.read() to return messages
that were sent from a previously opened port.