Read channel state from GPIO chips instead of tracking them in a vector
This commit makes the number_of_channels() member function dynamic, depending
on the number of HV switching boards attached to the control board. In
addition, the startup debug message now shows how many and which boards are
connected. In order to be considered valid, board addresses must start at 0
and increase by +1 for each additional board.

Closes #7.