Add several fixes related to microdrop bug 56
If the control board was measuring impedance, it was possible to
switch to a new resistor and not actually have enough time to get a
sample, which resulted in a reading of -1024. Now, when we don't
have sufficient time to get a reading, the fb/hv resistor index is
set to -1.

It still seems that there are occasinally problems with the
measurement during which the resistor switches. Perhaps we should
just ignore these?

See http://microfluidics.utoronto.ca/microdrop/ticket/56