Add hardware version define macros
These allow hardware specific functionality to be coded on the
Arduino. This commit also removed peak detector code, for which the
circuitry no longer exists post 1.2 (and was not used in earlier
versions).

Note that the scons script should eventually package hex files for
all existing versions and the software should automatically flash
the right version (or allow the user to choose).