Fix impedance calibration param fitting
Prior to this commit, we were solving the transfer function for `V2`, assuming
`R1=infinity`.  However, this results in `R1` in both the numerator and
denominator, producing a value of zero regardless of the values of the other
arguments.  As of this commit, we avoid this issue by specifying a *very large*
value for `R1` *(i.e., 10e45)*.

TODO Check with [Ryan][1] if this is OK.

[1]: mailto:ryan@fobel.net