Update amplifier gain/voltage when the actuation...
Update amplifier gain/voltage when the actuation exceeds 50% of the set voltage
tolerance. This should improve the capability of the actuation voltage to track
the target and reduce the number of "volage tolerance exceeded" errors. This
partially decouples the voltage tolerance (which signifies the maximum voltage
tolerance a user is willing to accept) from the amplifier gain update condition.