[NB] Re-raise firmware RuntimeError->FirmwareError
Prior to this commit, remote/firmware exceptions were raised as
`RuntimeError` instances with a message similar to:

    Error sending command 0xF4 (244). Return code=32.

As of this commit, several `DMFControlBoard` methods are wrapped with
the `remote_command` decorator.  For methods wrapped with the
`remote_command` decorator, `RuntimeError` exceptions raised by remote
control board firmware commands are caught and re-raised as more
specific `FirmwareError` exceptions, which include command code and
return code.

Notes
-----

 - Add `COMMAND_CODES_BY_NAME` and `NAMES_BY_COMMAND_CODE` mappings to
   `dmf_control_board_firmware` top-level module.
 - Add `RETURN_CODES_BY_NAME` and `NAMES_BY_RETURN_CODE` mappings to
   `dmf_control_board_firmware` top-level module.